+++ test/store.test.ts
@@ -1,70 +1,77 @@
-import { makeStore } from "../src/store"
-import { Listener } from "../src/types"
+import { makeStore } from "../src"
 
 describe("makeStore", () => {
-  const state = {
+  const store = makeStore({
     foo: 0,
     bar: 0
-  }
+  })
 
-  const store = makeStore(state)
+  beforeEach(() => {
+    store.set({ foo: 0, bar: 0 })
+  })
 
   describe(".state", () => {
     it("provides direct access to the state object", () => {
-      expect(store.state).toBe(state)
+      expect(store.state).toEqual({ foo: 0, bar: 0 })
     })
   })
 
   describe(".set", () => {
     it("accepts a dictionary of updates to the state", () => {
       store.set({ foo: 10 })
-      expect(state.foo).toEqual(10)
+      expect(store.state.foo).toEqual(10)
     })
 
     it("accepts a function that accepts the state and returns an update dictionary", () => {
-      const current = state.foo
+      const current = store.state.foo
+      expect(store.state.foo).toEqual(current)
       store.set(({ foo }) => ({ foo: foo + 1 }))
-      expect(state.foo).toEqual(current + 1)
+      expect(store.state.foo).toEqual(current + 1)
+    })
+
+    it("returns the updated state", () => {
+      const result = store.set({ foo: 1 })
+      expect(result).toEqual({ foo: 1, bar: 0 })
     })
   })
 
   describe(".subscribe", () => {
-    it("allows subscribing to updates to a specific property of the store", () => {
-      let fooChanges = 0
-      let barChanges = 0
+    it("accepts a listener callback that will be invoked when the store changes", () => {
+      const listener = jest.fn()
+      store.set({ foo: 0, bar: 0 })
+      store.subscribe(listener)
+      store.set({ foo: 1 })
+      store.unsubscribe(listener)
 
-      const fooListener = () => fooChanges++
-      const barListener = () => barChanges++
+      /* It should have been called exactly once */
+      expect(listener.mock.calls.length).toBe(1)
 
-      store.subscribe("foo", fooListener)
-      store.subscribe("bar", barListener)
+      /* The first argument should be the changes */
+      expect(listener.mock.calls[0][0]).toEqual({ foo: 1 })
+
+      /* The second argument should be the previous state */
+      expect(listener.mock.calls[0][1]).toEqual({ foo: 0, bar: 0 })
+    })
+
+    it("allows subscribing to updates to a store", () => {
+      const changeCounters = {
+        foo: 0,
+        bar: 0
+      }
+
+      const listener = (updates) => {
+        for (const prop in updates) changeCounters[prop]++
+      }
+
+      store.subscribe(listener)
 
       store.set(({ foo }) => ({ foo: foo + 1 }))
       store.set(({ foo, bar }) => ({ foo: foo + 1, bar: bar + 1 }))
 
-      store.unsubscribe("foo", fooListener)
-      store.unsubscribe("bar", barListener)
+      store.unsubscribe(listener)
 
-      expect(fooChanges).toEqual(2)
-      expect(barChanges).toEqual(1)
-    })
-
-    it("feeds the changed values to the listener callback", () => {
-      let newFoo: number
-      let prevFoo: number
-
-      const listener: Listener<number> = (newValue, prevValue) => {
-        newFoo = newValue
-        prevFoo = prevValue
-      }
-
-      store.set({ foo: 0 })
-      store.subscribe("foo", listener)
-      store.set({ foo: 1 })
-      store.unsubscribe("foo", listener)
-
-      expect(newFoo).toBe(1)
-      expect(prevFoo).toBe(0)
+      expect(changeCounters.foo).toEqual(2)
+      expect(changeCounters.bar).toEqual(1)
     })
   })
 })